Make use of acc.AddError(). Fix SAN-level stat aggregation issue
Moved to use acc.AddError() to collect the errors that occur, rather than reporting them on the console, as happened previously.

Also fixing an issue where the metrics pushed into InfluxDB were not including any of the data referred to in the tag. This meant that looking at a SAN as a whole, it wasn't able to aggregate the IO Usage and Storage Usage metrics to provide totals overall.